+++ build-install-test.bash
@@ -1,1 +1,2 @@
-gem build text-to-html.gemspec && sudo gem install text-to-html-0.0.0.gem && ruby test/test_text-to-html.rb
+gem build text-to-html.gemspec && sudo gem install text-to-html-0.0.1.gem && ruby test/test_text-to-html.rb
+